updated files according to review comments
