+++ react/week1/my-react-app/src/App.css
@@ -1,38 +1,21 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
+.todoitem{
   display: flex;
-  flex-direction: column;
-  align-items: center;
+  justify-content: space-evenly;
+  background-color: rgb(166, 166, 181);
+}
+.id{
+
+  color: brown;
+}
+.title{
+  color: rgb(13, 9, 55);
+}
+
+.time{
+  color: rgb(93, 27, 154);
+}
+.todoheader{
+  display: flex;
   justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+  background: none;
 }
